+++ src/main/java/com/tcdt/qlnvkhoach/repository/thongtriduyetydutoan/TtDuyetYDuToanRepositoryCustomImpl.java
@@ -1,34 +1,24 @@
 package com.tcdt.qlnvkhoach.repository.thongtriduyetydutoan;
 
 
-import com.tcdt.qlnvkhoach.entities.denghicapvonbonganh.KhDnCapVonBoNganhCt;
-import com.tcdt.qlnvkhoach.entities.denghicapvonbonganh.KhDnCapVonBoNganh_;
 import com.tcdt.qlnvkhoach.enums.NhapXuatHangTrangThaiEnum;
-import com.tcdt.qlnvkhoach.enums.Operator;
-import com.tcdt.qlnvkhoach.request.denghicapvonbonganh.KhDnCapVonBoNganhSearchRequest;
 import com.tcdt.qlnvkhoach.request.thongtriduyetydutoan.TtDuyetYDuToanSearchRequest;
-import com.tcdt.qlnvkhoach.response.denghicapvonbonganh.KhDnCapVonBoNganhSearchResponse;
 import com.tcdt.qlnvkhoach.response.thongtriduyetydutoan.TtDuyetYDuToanResponse;
-import com.tcdt.qlnvkhoach.table.catalog.QlnvDmDonvi;
-import com.tcdt.qlnvkhoach.table.catalog.QlnvDmDonvi_;
 import com.tcdt.qlnvkhoach.util.QueryUtils;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
-import javax.persistence.*;
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
+import javax.persistence.Tuple;
 import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Log4j2
@@ -58,13 +48,13 @@ public class TtDuyetYDuToanRepositoryCustomImpl implements TtDuyetYDuToanReposit
             TtDuyetYDuToanResponse kh = new TtDuyetYDuToanResponse();
             kh.setId(item.get("id", BigDecimal.class).longValue());
             kh.setSoThongTri(item.get("SO_THONG_TRI", String.class));
-            kh.setNam(item.get("NAM", Integer.class));
+            kh.setNam(item.get("NAM", BigDecimal.class).intValue());
 
             Timestamp x = item.get("NGAY_LAP", Timestamp.class);
             kh.setNgayLap(LocalDate.parse(x.toLocalDateTime().toLocalDate().format(DateTimeFormatter.ISO_LOCAL_DATE)));
 
             kh.setLyDoChi(item.get("LY_DO_CHI", String.class));
-            kh.setSoDnCapVon(item.get("SO_DN_CAP_VON", Integer.class));
+            kh.setSoDnCapVon(item.get("SO_DN_CAP_VON", BigDecimal.class));
             kh.setTenDvi(item.get("TEN_DVI", String.class));
             kh.setMaDvi(item.get("MA_DVI", String.class));
